Add README.md with instructions for remote Docker configuration
A README.md file has been added, detailing the process to configure a remote Docker host, which is necessary when running Test Containers as the Docker build process runs in an isolated environment. The README also includes steps to configure Test Containers to connect to a remote Docker environment. A reference to the README.md has been added to the movies.sln file.